Refactor migration files to check if tables exist before creating them

+++ database/migrations/2023_12_12_231212_create_pre_certificate_other_documents_table.php
@@ -13,28 +13,30 @@ class CreatePreCertificateOtherDocumentsTable extends Migration
      */
     public function up()
     {
-		Schema::create('pre_certificate_other_documents', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('pre_certificate_id');
-            $table->foreign('pre_certificate_id')
-                ->references('id')
-                ->on('pre_certificates')
-                ->onDelete('cascade');
-            $table->text('name')->nullable();
-            $table->text('link')->nullable();
-            $table->text('type')->nullable();
-            $table->text('size')->nullable();
-            $table->text('description')->nullable();
-            $table->timestamp('created_at')->useCurrent();
-            $table->uuid('created_by')->nullable();
+        if (!Schema::hasTable('pre_certificate_other_documents')) {
+            Schema::create('pre_certificate_other_documents', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('pre_certificate_id');
+                $table->foreign('pre_certificate_id')
+                    ->references('id')
+                    ->on('pre_certificates')
+                    ->onDelete('cascade');
+                $table->text('name')->nullable();
+                $table->text('link')->nullable();
+                $table->text('type')->nullable();
+                $table->text('size')->nullable();
+                $table->text('description')->nullable();
+                $table->timestamp('created_at')->useCurrent();
+                $table->uuid('created_by')->nullable();
                 $table->foreign('created_by')
                     ->references('id')
                     ->on('users')
                     ->onDelete('cascade');
-            $table->timestamp('updated_at')->useCurrent();
-            $table->text('type_document')->nullable();
-            $table->softDeletes();
-        });
+                $table->timestamp('updated_at')->useCurrent();
+                $table->text('type_document')->nullable();
+                $table->softDeletes();
+            });
+        }
     }
 
     /**
